# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleInput {
  projectId: number;
  postId: number;
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  date: string;
  time: string;
  channelIds?: number[];
}